# specs/001-add-go-backend/contracts/api.yaml
openapi: 3.1.0
info:
  title: Online-PPT Backend API
  version: 0.1.0
  description: REST API for user authentication and PPT record management.
servers:
  - url: http://localhost:8080/api/v1
paths:
  /auth/captcha:
    get:
      summary: Get captcha image for verification
      description: Generate a new captcha image and return base64-encoded image with captcha ID
      responses:
        '200':
          description: Captcha generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptchaResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/send-verification-code:
    post:
      summary: Send email verification code
      description: Validate captcha and send a 6-digit verification code to the provided email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendVerificationCodeRequest'
      responses:
        '200':
          description: Verification code sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendVerificationCodeResponse'
        '400':
          description: Invalid captcha or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limited - too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/register:
    post:
      summary: Register a new user with email verification code
      description: Complete user registration using email and verification code received via email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully and auto-login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error or invalid verification code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many verification attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      summary: Authenticate user and issue tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              description: HttpOnly cookie containing refresh token (if configured)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      summary: Revoke current session
      security:
        - AccessToken: []
      responses:
        '204':
          description: Logout successful
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      summary: Exchange refresh token for new access token
      responses:
        '200':
          description: New tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /ppts:
    get:
      summary: List PPT records for current user
      security:
        - AccessToken: []
      parameters:
        - name: q
          in: query
          description: Keyword to search in name/description/tags
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [created_at_desc, created_at_asc, name_asc]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PptRecordListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a PPT record
      security:
        - AccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PptRecordCreateRequest'
      responses:
        '201':
          description: Record created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PptRecordResponse'
        '400':
          description: Validation error or path outside allowed root
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Record with same path already exists for user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /ppts/{id}:
    get:
      summary: Fetch a single PPT record
      security:
        - AccessToken: []
      parameters:
        - $ref: '#/components/parameters/RecordId'
      responses:
        '200':
          description: Record details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PptRecordResponse'
        '404':
          description: Record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update a PPT record
      security:
        - AccessToken: []
      parameters:
        - $ref: '#/components/parameters/RecordId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PptRecordUpdateRequest'
      responses:
        '200':
          description: Record updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PptRecordResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a PPT record
      security:
        - AccessToken: []
      parameters:
        - $ref: '#/components/parameters/RecordId'
      responses:
        '204':
          description: Record deleted
        '404':
          description: Record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    AccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RecordId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    CaptchaResponse:
      type: object
      required: [captcha_id, image, expires_in]
      properties:
        captcha_id:
          type: string
          description: Unique identifier for the captcha
        image:
          type: string
          description: Base64-encoded captcha image
        expires_in:
          type: integer
          description: Captcha validity duration in seconds
    SendVerificationCodeRequest:
      type: object
      required: [email, captcha_id, captcha_code]
      properties:
        email:
          type: string
          format: email
          description: Email address to send verification code
        captcha_id:
          type: string
          description: Captcha ID obtained from /auth/captcha
        captcha_code:
          type: string
          description: User input for captcha verification
    SendVerificationCodeResponse:
      type: object
      required: [message, expires_in]
      properties:
        message:
          type: string
          description: Success message
        expires_in:
          type: integer
          description: Verification code validity duration in seconds
    RegisterRequest:
      type: object
      required: [email, password, email_code]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 10
        email_code:
          type: string
          description: 6-digit verification code received via email
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        accessToken:
          type: string
          description: Short-lived JWT access token
        expiresIn:
          type: integer
          description: Access token validity window in seconds
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        status:
          type: string
          enum: [active, locked, pending]
        lastLoginAt:
          type: [string, 'null']
          format: date-time
    PptRecordCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 120
          pattern: '^[A-Za-z0-9_-]+$'
          description: Display name and folder group identifier
        title:
          type: [string, 'null']
          maxLength: 255
          description: User-friendly display title supporting Chinese, English, numbers, spaces, and common punctuation
        description:
          type: [string, 'null']
          maxLength: 500
        tags:
          type: array
          items:
            type: string
          maxItems: 10
    PptRecordUpdateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 120
          pattern: '^[A-Za-z0-9_-]+$'
          description: If changed, server regenerates folder name
        title:
          type: [string, 'null']
          maxLength: 255
          description: User-friendly display title supporting Chinese, English, numbers, spaces, and common punctuation
        description:
          type: [string, 'null']
          maxLength: 500
        tags:
          type: array
          items:
            type: string
          maxItems: 10
    PptRecordResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        title:
          type: [string, 'null']
          description: User-friendly display title; if null, frontend should fallback to name
        groupName:
          type: string
        description:
          type: [string, 'null']
        relativePath:
          type: string
          description: Relative path in format presentations/<user-uuid>/<group>/slides
        canonicalPath:
          type: string
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        pathStatus:
          type: string
          enum: [valid, missing, outside_root]
      required: [id, name, relativePath, canonicalPath, createdAt, updatedAt, pathStatus]
    PptRecordListResponse:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/PptRecordResponse'
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: |
            Error code identifier:
            - invalid_request: Malformed request or validation error
            - invalid_captcha: Captcha verification failed or expired
            - invalid_code: Email verification code invalid or expired
            - email_exists: Email already registered
            - invalid_credentials: Login failed - invalid email or password
            - session_not_found: Refresh token invalid or expired
            - rate_limited: Too many requests, please retry later
            - too_many_attempts: Too many verification attempts, need new code
            - server_error: Internal server error
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          description: Optional additional error context